@@ -53,50 +53,52 @@
     "url": "https://github.com/caolan/async/issues"
   },
   "dependencies": {
-    "lodash": "^4.8.0"
+    "lodash": "^4.14.0"
   },
   "description": "Higher-order functions and common patterns for asynchronous code",
   "devDependencies": {
+    "babel-cli": "^6.16.0",
     "babel-core": "^6.3.26",
-    "babel-plugin-add-module-exports": "~0.1.2",
-    "babel-plugin-istanbul": "^1.0.3",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-plugin-istanbul": "^2.0.1",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.3.16",
     "babel-preset-es2015": "^6.3.13",
     "babelify": "^7.2.0",
-    "benchmark": "github:bestiejs/benchmark.js",
-    "bluebird": "^2.9.32",
+    "benchmark": "^2.1.1",
+    "bluebird": "^3.4.6",
     "chai": "^3.1.0",
-    "cheerio": "^0.20.0",
+    "cheerio": "^0.22.0",
     "coveralls": "^2.11.2",
     "es6-promise": "^2.3.0",
-    "eslint": "^2.11.1",
+    "eslint": "^2.13.1",
     "fs-extra": "^0.26.7",
     "gh-pages-deploy": "^0.4.2",
     "jsdoc": "^3.4.0",
-    "karma": "^0.13.2",
-    "karma-browserify": "^4.2.1",
-    "karma-firefox-launcher": "^0.1.6",
-    "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.0.2",
-    "mocha": "^2.2.5",
+    "karma": "^1.3.0",
+    "karma-browserify": "^5.1.0",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-mocha": "^1.2.0",
+    "karma-mocha-reporter": "^2.2.0",
+    "mocha": "^3.1.2",
     "native-promise-only": "^0.8.0-a",
     "nyc": "^7.0.0",
     "recursive-readdir": "^1.3.0",
     "rimraf": "^2.5.0",
-    "rollup": "^0.25.0",
-    "rollup-plugin-node-resolve": "^1.5.0",
-    "rollup-plugin-npm": "~1.3.0",
+    "rollup": "^0.36.3",
+    "rollup-plugin-node-resolve": "^2.0.0",
+    "rollup-plugin-npm": "^2.0.0",
     "rsvp": "^3.0.18",
     "semver": "^4.3.6",
-    "uglify-js": "~2.4.0",
+    "uglify-js": "~2.7.3",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
+    "watchify": "^3.7.0",
     "yargs": "~3.9.1"
   },
   "directories": {},
   "dist": {
-    "shasum": "b709cc0280a9c36f09f4536be823c838a9049e25",
-    "tarball": "https://registry.npmjs.org/async/-/async-2.0.1.tgz"
+    "shasum": "2d2160c7788032e4dd6cbe2502f1f9a2c8f6cde4",
+    "tarball": "https://registry.npmjs.org/async/-/async-2.1.4.tgz"
   },
   "gh-pages-deploy": {
     "staticpath": "docs"
