@@ -58,33 +58,38 @@
   "dependencies": {
     "asynckit": "^0.4.0",
     "combined-stream": "^1.0.5",
-    "mime-types": "^2.1.11"
+    "mime-types": "^2.1.12"
   },
   "description": "A library to create readable \"multipart/form-data\" streams. Can be used to submit forms and file uploads to other web applications.",
   "devDependencies": {
-    "coveralls": "^2.11.13",
-    "cross-spawn": "^4.0.0",
-    "eslint": "^3.5.0",
+    "browserify": "^13.1.1",
+    "browserify-istanbul": "^2.0.0",
+    "coveralls": "^2.11.14",
+    "cross-spawn": "^4.0.2",
+    "eslint": "^3.9.1",
     "fake": "^0.2.2",
     "far": "^0.0.7",
     "formidable": "^1.0.17",
     "in-publish": "^2.0.0",
     "is-node-modern": "^1.0.0",
     "istanbul": "^0.4.5",
+    "obake": "^0.1.2",
+    "phantomjs-prebuilt": "^2.1.13",
     "pkgfiles": "^2.3.0",
     "pre-commit": "^1.1.3",
-    "request": "^2.74.0",
-    "rimraf": "^2.5.4"
+    "request": "2.76.0",
+    "rimraf": "^2.5.4",
+    "tape": "^4.6.2"
   },
   "directories": {},
   "dist": {
-    "shasum": "6f0aebadcc5da16c13e1ecc11137d85f9b883b25",
-    "tarball": "https://registry.npmjs.org/form-data/-/form-data-2.0.0.tgz"
+    "shasum": "89c3534008b97eada4cbb157d58f6f5df025eae4",
+    "tarball": "https://registry.npmjs.org/form-data/-/form-data-2.1.2.tgz"
   },
   "engines": {
     "node": ">= 0.12"
   },
-  "gitHead": "652b16ff5b9077bdf65eb66b67286c823c2a1040",
+  "gitHead": "03444d21961a7a44cdc2eae11ee3630f6969023d",
   "homepage": "https://github.com/form-data/form-data#readme",
   "license": "MIT",
   "main": "./lib/form_data",
