@@ -2,46 +2,50 @@
   "_args": [
     [
       {
-        "raw": "jsonpointer@2.0.0",
+        "raw": "jsonpointer@^4.0.0",
         "scope": null,
         "escapedName": "jsonpointer",
         "name": "jsonpointer",
-        "rawSpec": "2.0.0",
-        "spec": "2.0.0",
-        "type": "version"
+        "rawSpec": "^4.0.0",
+        "spec": ">=4.0.0 <5.0.0",
+        "type": "range"
       },
-      "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-116827-2000-1gmttnr/node_modules/atom-beautify/node_modules/is-my-json-valid"
+      "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-11716-65523-1476sqp/node_modules/atom-beautify/node_modules/is-my-json-valid"
     ]
   ],
-  "_from": "jsonpointer@2.0.0",
-  "_id": "jsonpointer@2.0.0",
+  "_from": "jsonpointer@>=4.0.0 <5.0.0",
+  "_id": "jsonpointer@4.0.1",
   "_inCache": true,
   "_installable": true,
   "_location": "/atom-beautify/jsonpointer",
-  "_nodeVersion": "0.10.36",
+  "_nodeVersion": "6.9.1",
+  "_npmOperationalInternal": {
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/jsonpointer-4.0.1.tgz_1482326391770_0.6748844815883785"
+  },
   "_npmUser": {
     "name": "marcbachmann",
     "email": "marc.brookman@gmail.com"
   },
-  "_npmVersion": "2.10.1",
+  "_npmVersion": "3.10.8",
   "_phantomChildren": {},
   "_requested": {
-    "raw": "jsonpointer@2.0.0",
+    "raw": "jsonpointer@^4.0.0",
     "scope": null,
     "escapedName": "jsonpointer",
     "name": "jsonpointer",
-    "rawSpec": "2.0.0",
-    "spec": "2.0.0",
-    "type": "version"
+    "rawSpec": "^4.0.0",
+    "spec": ">=4.0.0 <5.0.0",
+    "type": "range"
   },
   "_requiredBy": [
     "/atom-beautify/is-my-json-valid"
   ],
-  "_resolved": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-2.0.0.tgz",
-  "_shasum": "3af1dd20fe85463910d469a385e33017d2a030d9",
+  "_resolved": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-4.0.1.tgz",
+  "_shasum": "4fd92cb34e0e9db3c89c8622ecf51f9b978c6cb9",
   "_shrinkwrap": null,
-  "_spec": "jsonpointer@2.0.0",
-  "_where": "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-116827-2000-1gmttnr/node_modules/atom-beautify/node_modules/is-my-json-valid",
+  "_spec": "jsonpointer@^4.0.0",
+  "_where": "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-11716-65523-1476sqp/node_modules/atom-beautify/node_modules/is-my-json-valid",
   "author": {
     "name": "Jan Lehnardt",
     "email": "jan@apache.org"
@@ -53,20 +57,29 @@
     {
       "name": "Joe Hildebrand",
       "email": "joe-github@cursive.net"
+    },
+    {
+      "name": "Marc Bachmann",
+      "email": "marc.brookman@gmail.com"
     }
   ],
   "dependencies": {},
   "description": "Simple JSON Addressing.",
-  "devDependencies": {},
+  "devDependencies": {
+    "standard": "^5.3.1"
+  },
   "directories": {},
   "dist": {
-    "shasum": "3af1dd20fe85463910d469a385e33017d2a030d9",
-    "tarball": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-2.0.0.tgz"
+    "shasum": "4fd92cb34e0e9db3c89c8622ecf51f9b978c6cb9",
+    "tarball": "https://registry.npmjs.org/jsonpointer/-/jsonpointer-4.0.1.tgz"
   },
   "engines": {
-    "node": ">=0.6.0"
+    "node": ">=0.10.0"
   },
-  "gitHead": "26ea4a5c0fcb6d9a2e87f733403791dd05637af8",
+  "files": [
+    "jsonpointer.js"
+  ],
+  "gitHead": "37c73aecd5f192a00cd79c0ebbbb2034b91bcfd0",
   "homepage": "https://github.com/janl/node-jsonpointer#readme",
   "license": "MIT",
   "main": "./jsonpointer",
@@ -88,7 +101,7 @@
     "url": "git+ssh://git@github.com/janl/node-jsonpointer.git"
   },
   "scripts": {
-    "test": "node test.js"
+    "test": "standard && node test.js"
   },
   "tags": [
     "util",
@@ -96,5 +109,5 @@
     "util",
     "utility"
   ],
-  "version": "2.0.0"
+  "version": "4.0.1"
 }
