@@ -10,18 +10,18 @@
         "spec": ">=2.1.7 <2.2.0",
         "type": "range"
       },
-      "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-116827-2000-1gmttnr/node_modules/atom-beautify/node_modules/request"
+      "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-11716-65523-1476sqp/node_modules/atom-beautify/node_modules/request"
     ]
   ],
   "_from": "mime-types@>=2.1.7 <2.2.0",
-  "_id": "mime-types@2.1.12",
+  "_id": "mime-types@2.1.14",
   "_inCache": true,
   "_installable": true,
   "_location": "/atom-beautify/mime-types",
-  "_nodeVersion": "4.5.0",
+  "_nodeVersion": "4.6.1",
   "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/mime-types-2.1.12.tgz_1474237415119_0.03028594213537872"
+    "host": "packages-18-east.internal.npmjs.com",
+    "tmp": "tmp/mime-types-2.1.14.tgz_1484458141358_0.7563668976072222"
   },
   "_npmUser": {
     "name": "dougwilson",
@@ -42,11 +42,11 @@
     "/atom-beautify/form-data",
     "/atom-beautify/request"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.12.tgz",
-  "_shasum": "152ba256777020dd4663f54c2e7bc26381e71729",
+  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.14.tgz",
+  "_shasum": "f7ef7d97583fcaf3b7d282b6f8b5679dab1e94ee",
   "_shrinkwrap": null,
   "_spec": "mime-types@~2.1.7",
-  "_where": "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-116827-2000-1gmttnr/node_modules/atom-beautify/node_modules/request",
+  "_where": "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-11716-65523-1476sqp/node_modules/atom-beautify/node_modules/request",
   "bugs": {
     "url": "https://github.com/jshttp/mime-types/issues"
   },
@@ -67,21 +67,21 @@
     }
   ],
   "dependencies": {
-    "mime-db": "~1.24.0"
+    "mime-db": "~1.26.0"
   },
   "description": "The ultimate javascript content-type utility.",
   "devDependencies": {
-    "eslint": "3.5.0",
-    "eslint-config-standard": "6.0.1",
-    "eslint-plugin-promise": "2.0.1",
-    "eslint-plugin-standard": "2.0.0",
+    "eslint": "3.13.1",
+    "eslint-config-standard": "6.2.1",
+    "eslint-plugin-promise": "3.4.0",
+    "eslint-plugin-standard": "2.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5"
   },
   "directories": {},
   "dist": {
-    "shasum": "152ba256777020dd4663f54c2e7bc26381e71729",
-    "tarball": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.12.tgz"
+    "shasum": "f7ef7d97583fcaf3b7d282b6f8b5679dab1e94ee",
+    "tarball": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.14.tgz"
   },
   "engines": {
     "node": ">= 0.6"
@@ -91,7 +91,7 @@
     "LICENSE",
     "index.js"
   ],
-  "gitHead": "7193a9094e2efe31da93988350bb0b32ab18b1ea",
+  "gitHead": "d7d15e50fe6b3a2da79d855015d25efa50e9f157",
   "homepage": "https://github.com/jshttp/mime-types#readme",
   "keywords": [
     "mime",
@@ -120,10 +120,10 @@
     "url": "git+https://github.com/jshttp/mime-types.git"
   },
   "scripts": {
-    "lint": "eslint **/*.js",
+    "lint": "eslint .",
     "test": "mocha --reporter spec test/test.js",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot test/test.js",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter dot test/test.js"
   },
-  "version": "2.1.12"
+  "version": "2.1.14"
 }
