@@ -59,26 +59,27 @@
   "dependencies": {},
   "description": "A querystring parser that supports nesting and arrays, with a depth limit",
   "devDependencies": {
-    "@ljharb/eslint-config": "^6.0.0",
-    "browserify": "^13.0.1",
+    "@ljharb/eslint-config": "^8.0.0",
+    "browserify": "^13.1.0",
     "covert": "^1.1.0",
-    "eslint": "^3.1.0",
+    "eslint": "^3.8.0",
     "evalmd": "^0.0.17",
     "iconv-lite": "^0.4.13",
     "mkdirp": "^0.5.1",
     "parallelshell": "^2.0.0",
     "qs-iconv": "^1.0.3",
-    "tape": "^4.6.0"
+    "safe-publish-latest": "^1.1.1",
+    "tape": "^4.6.2"
   },
   "directories": {},
   "dist": {
-    "shasum": "ce03c5ff0935bc1d9d69a9f14cbd18e568d67625",
-    "tarball": "https://registry.npmjs.org/qs/-/qs-6.2.1.tgz"
+    "shasum": "f403b264f23bc01228c74131b407f18d5ea5d442",
+    "tarball": "https://registry.npmjs.org/qs/-/qs-6.3.0.tgz"
   },
   "engines": {
     "node": ">=0.6"
   },
-  "gitHead": "335f839142e6c2c69f5302c4940d92acb0e77561",
+  "gitHead": "8aa9c26f90335b5483a4f456dea9acbada8a881c",
   "homepage": "https://github.com/ljharb/qs",
   "keywords": [
     "querystring",
