@@ -10,18 +10,18 @@
         "spec": ">=2.0.0 <3.0.0",
         "type": "range"
       },
-      "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-116827-2000-1gmttnr/node_modules/atom-beautify/node_modules/universal-analytics"
+      "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-11716-65523-1476sqp/node_modules/atom-beautify/node_modules/universal-analytics"
     ]
   ],
   "_from": "request@>=2.0.0 <3.0.0",
-  "_id": "request@2.75.0",
+  "_id": "request@2.79.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/atom-beautify/request",
-  "_nodeVersion": "6.5.0",
+  "_nodeVersion": "6.3.1",
   "_npmOperationalInternal": {
-    "host": "packages-16-east.internal.npmjs.com",
-    "tmp": "tmp/request-2.75.0.tgz_1474151606844_0.8052814984694123"
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/request-2.79.0.tgz_1479489666177_0.7806831058114767"
   },
   "_npmUser": {
     "name": "simov",
@@ -41,11 +41,11 @@
   "_requiredBy": [
     "/atom-beautify/universal-analytics"
   ],
-  "_resolved": "https://registry.npmjs.org/request/-/request-2.75.0.tgz",
-  "_shasum": "d2b8268a286da13eaa5d01adf5d18cc90f657d93",
+  "_resolved": "https://registry.npmjs.org/request/-/request-2.79.0.tgz",
+  "_shasum": "4dfe5bf6be8b8cdc37fcf93e04b65577722710de",
   "_shrinkwrap": null,
   "_spec": "request@2.x",
-  "_where": "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-116827-2000-1gmttnr/node_modules/atom-beautify/node_modules/universal-analytics",
+  "_where": "/private/var/folders/lq/w_0sst250637p0vm68llylyr0000gn/T/apm-install-dir-11716-65523-1476sqp/node_modules/atom-beautify/node_modules/universal-analytics",
   "author": {
     "name": "Mikeal Rogers",
     "email": "mikeal.rogers@gmail.com"
@@ -56,12 +56,11 @@
   "dependencies": {
     "aws-sign2": "~0.6.0",
     "aws4": "^1.2.1",
-    "bl": "~1.1.2",
     "caseless": "~0.11.0",
     "combined-stream": "~1.0.5",
     "extend": "~3.0.0",
     "forever-agent": "~0.6.1",
-    "form-data": "~2.0.0",
+    "form-data": "~2.1.1",
     "har-validator": "~2.0.6",
     "hawk": "~3.1.3",
     "http-signature": "~1.1.0",
@@ -69,12 +68,12 @@
     "isstream": "~0.1.2",
     "json-stringify-safe": "~5.0.1",
     "mime-types": "~2.1.7",
-    "node-uuid": "~1.4.7",
     "oauth-sign": "~0.8.1",
-    "qs": "~6.2.0",
+    "qs": "~6.3.0",
     "stringstream": "~0.0.4",
     "tough-cookie": "~2.3.0",
-    "tunnel-agent": "~0.4.1"
+    "tunnel-agent": "~0.4.1",
+    "uuid": "^3.0.0"
   },
   "description": "Simplified HTTP request client.",
   "devDependencies": {
@@ -97,17 +96,22 @@
     "rimraf": "^2.2.8",
     "server-destroy": "^1.0.1",
     "tape": "^4.6.0",
-    "taper": "^0.4.0"
+    "taper": "^0.5.0"
   },
   "directories": {},
   "dist": {
-    "shasum": "d2b8268a286da13eaa5d01adf5d18cc90f657d93",
-    "tarball": "https://registry.npmjs.org/request/-/request-2.75.0.tgz"
+    "shasum": "4dfe5bf6be8b8cdc37fcf93e04b65577722710de",
+    "tarball": "https://registry.npmjs.org/request/-/request-2.79.0.tgz"
   },
   "engines": {
-    "node": ">=0.8.0"
+    "node": ">= 4"
   },
-  "gitHead": "e9f09c2832073858d6d988ba82a2895f36efa92d",
+  "files": [
+    "lib/",
+    "index.js",
+    "request.js"
+  ],
+  "gitHead": "ff729c6f1a87237060d075908563ce13386395ac",
   "greenkeeper": {
     "ignore": [
       "eslint",
@@ -156,5 +160,5 @@
     "util",
     "utility"
   ],
-  "version": "2.75.0"
+  "version": "2.79.0"
 }
