@@ -120,33 +125,106 @@
     {
       "name": "David Glasser",
       "email": "glasser@davidglasser.net"
+    },
+    {
+      "name": "Simon Lydell",
+      "email": "simon.lydell@gmail.com"
+    },
+    {
+      "name": "Jmeas Smith",
+      "email": "jellyes2@gmail.com"
+    },
+    {
+      "name": "Michael Z Goddard",
+      "email": "mzgoddard@gmail.com"
+    },
+    {
+      "name": "azu",
+      "email": "azu@users.noreply.github.com"
+    },
+    {
+      "name": "John Gozde",
+      "email": "john@gozde.ca"
+    },
+    {
+      "name": "Adam Kirkton",
+      "email": "akirkton@truefitinnovation.com"
+    },
+    {
+      "name": "Chris Montgomery",
+      "email": "christopher.montgomery@dowjones.com"
+    },
+    {
+      "name": "J. Ryan Stinnett",
+      "email": "jryans@gmail.com"
+    },
+    {
+      "name": "Jack Herrington",
+      "email": "jherrington@walmartlabs.com"
+    },
+    {
+      "name": "Chris Truter",
+      "email": "jeffpalentine@gmail.com"
+    },
+    {
+      "name": "Daniel Espeset",
+      "email": "daniel@danielespeset.com"
+    },
+    {
+      "name": "Jamie Wong",
+      "email": "jamie.lf.wong@gmail.com"
+    },
+    {
+      "name": "Eddy Bruël",
+      "email": "ejpbruel@mozilla.com"
+    },
+    {
+      "name": "Hawken Rives",
+      "email": "hawkrives@gmail.com"
+    },
+    {
+      "name": "Gilad Peleg",
+      "email": "giladp007@gmail.com"
+    },
+    {
+      "name": "djchie",
+      "email": "djchie.dev@gmail.com"
+    },
+    {
+      "name": "Gary Ye",
+      "email": "garysye@gmail.com"
+    },
+    {
+      "name": "Nicolas Lalevée",
+      "email": "nicolas.lalevee@hibnet.org"
     }
   ],
-  "dependencies": {
-    "amdefine": ">=0.0.4"
-  },
+  "dependencies": {},
   "description": "Generates and consumes source maps",
   "devDependencies": {
-    "dryice": ">=0.4.8"
+    "doctoc": "^0.15.0",
+    "webpack": "^1.12.0"
   },
-  "directories": {
-    "lib": "./lib"
-  },
+  "directories": {},
   "dist": {
-    "shasum": "c8b6c167797ba4740a8ea33252162ff08591b266",
-    "tarball": "https://registry.npmjs.org/source-map/-/source-map-0.1.32.tgz"
+    "shasum": "75ce38f52bf0733c5a7f0c118d81334a2bb5f412",
+    "tarball": "https://registry.npmjs.org/source-map/-/source-map-0.5.6.tgz"
   },
   "engines": {
-    "node": ">=0.8.0"
+    "node": ">=0.10.0"
   },
-  "homepage": "https://github.com/mozilla/source-map",
-  "licenses": [
-    {
-      "type": "BSD",
-      "url": "http://opensource.org/licenses/BSD-3-Clause"
-    }
+  "files": [
+    "source-map.js",
+    "lib/",
+    "dist/source-map.debug.js",
+    "dist/source-map.js",
+    "dist/source-map.min.js",
+    "dist/source-map.min.js.map"
   ],
-  "main": "./lib/source-map.js",
+  "gitHead": "aa0398ced67beebea34f0d36f766505656c344f6",
+  "homepage": "https://github.com/mozilla/source-map",
+  "license": "BSD-3-Clause",
+  "main": "./source-map.js",
   "maintainers": [
     {
       "name": "mozilla-devtools",
