@@ -55,23 +60,24 @@
   "dependencies": {},
   "description": "Port of TweetNaCl cryptographic library to JavaScript",
   "devDependencies": {
-    "browserify": "^10.1.3",
-    "eslint": "^1.4.3",
-    "faucet": "0.0.1",
+    "browserify": "^13.0.0",
+    "eslint": "^2.2.0",
+    "faucet": "^0.0.1",
     "tap-browser-color": "^0.1.2",
-    "tape": "^4.0.0",
-    "testling": "^1.7.1",
-    "uglify-js": "^2.4.21"
+    "tape": "^4.4.0",
+    "tape-run": "^2.1.3",
+    "tweetnacl-util": "^0.13.3",
+    "uglify-js": "^2.6.1"
   },
   "directories": {
     "test": "test"
   },
   "dist": {
-    "shasum": "d628b56f3bcc3d5ae74ba9d4c1a704def5ab4b56",
-    "tarball": "https://registry.npmjs.org/tweetnacl/-/tweetnacl-0.13.3.tgz"
+    "shasum": "5ae68177f192d4456269d108afa93ff8743f4f64",
+    "tarball": "https://registry.npmjs.org/tweetnacl/-/tweetnacl-0.14.5.tgz"
   },
-  "gitHead": "2bb422cb707fba4a5ec9654688564a4fb861b068",
-  "homepage": "https://dchest.github.io/tweetnacl-js",
+  "gitHead": "cce829e473b1ae299a9373b5140c713ee88f577f",
+  "homepage": "https://tweetnacl.js.org",
   "keywords": [
     "crypto",
     "cryptography",
