@@ -1,7 +1,6 @@
 # There's an extra empty line before the closing of the class, that should irk
 # the Rubocop default config.
 class EmptyLinesAroundClassBody
-  def noop
-  end
+  def noop; end
 
 end
