@@ -1,8 +1,8 @@
-var i = 1;
-var run = setInterval(function() {
-  console.log("line " + i);
+let i = 1;
+const run = setInterval(() => {
+  console.log(`line ${i}`);
   i++;
-  if (i == 20) {
+  if (i === 20) {
     stop();
   }
 }, 1000);
