@@ -1,8 +1,8 @@
 process.stdin.setEncoding('utf8');
 
-process.stdin.on('readable', function() {
-  var chunk = process.stdin.read();
-   if (chunk !== null) {
-     console.log('TEST: ' + chunk);
-   }
+process.stdin.on('readable', () => {
+  const chunk = process.stdin.read();
+  if (chunk) {
+    console.log(`TEST: ${chunk}`);
+  }
 });
