@@ -1,6 +1,6 @@
 process.stdin.resume();
 process.stdin.setEncoding('utf8');
 
-process.stdin.on('end', function() {
-  console.log("stdin terminated");
+process.stdin.on('end', () => {
+  console.log('stdin terminated');
 });
